+++ river/models/__init__.py
@@ -5,5 +5,5 @@ from .state import *
 from .transition import *
 from .proceeding_meta import *
 from .proceeding import *
-from .proceeding_track import *
+# from .proceeding_track import *
 from .handler import *