+++ river/signals.py
@@ -1,5 +1,6 @@
-from django.dispatch import Signal
 import logging
+
+from django.dispatch import Signal
 
 __author__ = 'ahmetdal'
 
@@ -9,8 +10,8 @@ post_final = Signal(providing_args=["workflow_object", "field"])
 pre_transition = Signal(providing_args=["workflow_object", "field", "source_state", "destination_state", "proceeding"])
 post_transition = Signal(providing_args=["workflow_object", "field", "source_state", "destination_state", "proceeding"])
 
-pre_proceed = Signal(providing_args=["workflow_object", "field", "proceeding", "track"])
-post_proceed = Signal(providing_args=["workflow_object", "field", "proceeding", "track"])
+pre_proceed = Signal(providing_args=["workflow_object", "field", "proceeding"])
+post_proceed = Signal(providing_args=["workflow_object", "field", "proceeding"])
 
 LOGGER = logging.getLogger(__name__)
 
@@ -33,9 +34,12 @@ class TransitionSignal(object):
                 proceeding=self.proceeding,
             )
             LOGGER.debug("Pre transition signal IS sent for workflow object %s for field %s for transition %s -> %s" % (
-                self.workflow_object, self.field, self.proceeding.meta.transition.source_state.label, self.proceeding.meta.transition.destination_state.label))
+                self.workflow_object, self.field, self.proceeding.meta.transition.source_state.label,
+                self.proceeding.meta.transition.destination_state.label))
         else:
-            LOGGER.debug("Pre transition signal IS NOT sent for workflow object %s for field %s. Although proceeding is occurred, not transition is occurred." % (self.workflow_object, self.field))
+            LOGGER.debug(
+                "Pre transition signal IS NOT sent for workflow object %s for field %s. Although proceeding is occurred, not transition is occurred." % (
+                self.workflow_object, self.field))
 
     def __exit__(self, type, value, traceback):
         if self.status:
@@ -47,18 +51,21 @@ class TransitionSignal(object):
                 destination_state=self.proceeding.meta.transition.destination_state,
                 proceeding=self.proceeding,
             )
-            LOGGER.debug("Post transition signal IS sent for workflow object %s for field %s for transition %s -> %s" % (
-                self.workflow_object, self.field, self.proceeding.meta.transition.source_state.label, self.proceeding.meta.transition.destination_state.label))
+            LOGGER.debug(
+                "Post transition signal IS sent for workflow object %s for field %s for transition %s -> %s" % (
+                    self.workflow_object, self.field, self.proceeding.meta.transition.source_state.label,
+                    self.proceeding.meta.transition.destination_state.label))
         else:
-            LOGGER.debug("Post transition signal IS NOT sent for workflow object %s for field %s. Although proceeding is occurred, not transition is occurred." % (self.workflow_object, self.field))
+            LOGGER.debug(
+                "Post transition signal IS NOT sent for workflow object %s for field %s. Although proceeding is occurred, not transition is occurred." % (
+                self.workflow_object, self.field))
 
 
 class ProceedingSignal(object):
-    def __init__(self, workflow_object, field, proceeding, track):
+    def __init__(self, workflow_object, field, proceeding):
         self.workflow_object = workflow_object
         self.field = field
         self.proceeding = proceeding
-        self.track = track
 
     def __enter__(self):
         pre_proceed.send(
@@ -66,10 +73,10 @@ class ProceedingSignal(object):
             workflow_object=self.workflow_object,
             field=self.field,
             proceeding=self.proceeding,
-            track=self.track
         )
         LOGGER.debug("Pre proceeding signal IS sent for workflow object %s for field %s for transition %s -> %s" % (
-            self.workflow_object, self.field, self.proceeding.meta.transition.source_state.label, self.proceeding.meta.transition.destination_state.label))
+            self.workflow_object, self.field, self.proceeding.meta.transition.source_state.label,
+            self.proceeding.meta.transition.destination_state.label))
 
     def __exit__(self, type, value, traceback):
         post_proceed.send(
@@ -77,10 +84,10 @@ class ProceedingSignal(object):
             workflow_object=self.workflow_object,
             field=self.field,
             proceeding=self.proceeding,
-            track=self.track
         )
         LOGGER.debug("Post proceeding signal IS sent for workflow object %s for field %s for transition %s -> %s" % (
-            self.workflow_object, self.field, self.proceeding.meta.transition.source_state.label, self.proceeding.meta.transition.destination_state.label))
+            self.workflow_object, self.field, self.proceeding.meta.transition.source_state.label,
+            self.proceeding.meta.transition.destination_state.label))
 
 
 class FinalSignal(object):
@@ -96,7 +103,8 @@ class FinalSignal(object):
                 workflow_object=self.workflow_object,
                 field=self.field
             )
-            LOGGER.debug("Pre final signal IS sent for workflow object %s for field %s for final state %s" % (self.workflow_object, self.field, getattr(self.workflow_object, self.field).label))
+            LOGGER.debug("Pre final signal IS sent for workflow object %s for field %s for final state %s" % (
+            self.workflow_object, self.field, getattr(self.workflow_object, self.field).label))
 
     def __exit__(self, type, value, traceback):
         if self.status:
@@ -105,4 +113,5 @@ class FinalSignal(object):
                 workflow_object=self.workflow_object,
                 field=self.field,
             )
-            LOGGER.debug("Post final signal IS sent for workflow object %s for field %s for final state %s" % (self.workflow_object, self.field, getattr(self.workflow_object, self.field).label))
+            LOGGER.debug("Post final signal IS sent for workflow object %s for field %s for final state %s" % (
+            self.workflow_object, self.field, getattr(self.workflow_object, self.field).label))