+++ river/tests/handlers/backends/test__database_handler_backend.py
@@ -1,9 +1,9 @@
+from river.config import app_config
 from river.handlers.backends.loader import load_handler_backend
 from river.handlers.transition import PostTransitionHandler
 from river.models.handler import Handler
-from river.config import app_config
 from river.services.object import ObjectService
-from river.tests.services.proceeding_service_based_test import ProceedingServiceBasedTest
+from river.tests.base_test import BaseTestCase
 
 __author__ = 'ahmetdal'
 
@@ -12,10 +12,10 @@ def test_handler(*args, **kwargs):
     pass
 
 
-class test_DatabaseHandlerBackend(ProceedingServiceBasedTest):
+class test_DatabaseHandlerBackend(BaseTestCase):
     def setUp(self):
         super(test_DatabaseHandlerBackend, self).setUp()
-
+        self.initialize_normal_scenario()
         app_config.HANDLER_BACKEND_CLASS = 'river.handlers.backends.database.DatabaseHandlerBackend'
         self.handler_backend = load_handler_backend()
         self.handler_backend.handlers = {}
@@ -24,12 +24,17 @@ class test_DatabaseHandlerBackend(ProceedingServiceBasedTest):
 
     def test_register(self):
         self.assertEqual(0, Handler.objects.count())
-        self.assertFalse('%s.%s_object%sfield%s' % (PostTransitionHandler.__module__, PostTransitionHandler.__name__, self.objects[1].pk, 'my_field') in self.handler_backend.handlers)
+        self.assertFalse('%s.%s_object%sfield%s' % (
+            PostTransitionHandler.__module__, PostTransitionHandler.__name__, self.objects[1].pk,
+            'my_field') in self.handler_backend.handlers)
 
         self.handler_backend.register(PostTransitionHandler, test_handler, self.objects[1], 'my_field')
 
-        self.assertEqual(1, Handler.objects.filter(hash='%s.%s_object%sfield%s' % (PostTransitionHandler.__module__, PostTransitionHandler.__name__, self.objects[1].pk, 'my_field')).count())
-        self.assertTrue('%s.%s_object%sfield%s' % (PostTransitionHandler.__module__, PostTransitionHandler.__name__, self.objects[1].pk, 'my_field') in self.handler_backend.handlers)
+        self.assertEqual(1, Handler.objects.filter(hash='%s.%s_object%sfield%s' % (
+            PostTransitionHandler.__module__, PostTransitionHandler.__name__, self.objects[1].pk, 'my_field')).count())
+        self.assertTrue('%s.%s_object%sfield%s' % (
+            PostTransitionHandler.__module__, PostTransitionHandler.__name__, self.objects[1].pk,
+            'my_field') in self.handler_backend.handlers)
         self.assertEqual(test_handler.__name__, list(self.handler_backend.handlers.values())[0].__name__)
 
     def test_register_in_multiprocessing(self):
@@ -63,7 +68,8 @@ class test_DatabaseHandlerBackend(ProceedingServiceBasedTest):
         from multiprocessing import Process, Queue
 
         Handler.objects.update_or_create(
-            hash='%s.%s_object%sfield%s' % (PostTransitionHandler.__module__, PostTransitionHandler.__name__, self.objects[1].pk, 'my_field'),
+            hash='%s.%s_object%sfield%s' % (
+                PostTransitionHandler.__module__, PostTransitionHandler.__name__, self.objects[1].pk, 'my_field'),
             defaults={
                 'method': '%s.%s' % (test_handler.__module__, test_handler.__name__),
                 'handler_cls': '%s.%s' % (PostTransitionHandler.__module__, PostTransitionHandler.__name__),