+++ river/tests/services/test__object_service.py
@@ -1,14 +1,16 @@
 from river.models.proceeding import Proceeding
 from river.services.object import ObjectService
 from river.services.state import StateService
-from river.tests.services.proceeding_service_based_test import ProceedingServiceBasedTest
+from river.tests.base_test import BaseTestCase
 
 __author__ = 'ahmetdal'
 
 
 # noinspection PyPep8Naming
-class test_ObjectService(ProceedingServiceBasedTest):
+class test_ObjectService(BaseTestCase):
     def test_init(self):
+        self.initialize_normal_scenario()
+
         ObjectService.register_object(self.objects[0], self.field)
         ObjectService.register_object(self.objects[1], self.field)
 
@@ -18,6 +20,8 @@ class test_ObjectService(ProceedingServiceBasedTest):
         self.assertEqual(18, Proceeding.objects.count())
 
     def test_get_objects_waiting_for_approval_for_user(self):
+        self.initialize_normal_scenario()
+
         on_approval_objects = ObjectService.get_objects_waiting_for_approval(self.content_type, self.field, self.user1)
         self.assertEqual(2, on_approval_objects.count())
         self.assertEqual(self.objects[0], on_approval_objects[0])