+++ docs/changelog.rst
@@ -3,6 +3,12 @@
 Change Logs
 ===========
 
+0.8.0 (Dev)
+-----------
+
+* **Deprecation** - ProceedingTrack is removed. ProceedingTracks were being used to keep any transaction track to handle even circular one. This was a workaround. So, it can be handled with Proceeding now by cloning them if there is circle. ProceedingTracks was just causing confusion. To fix this, ProceedingTrack model and its functions are removed from django-river.
+* **Improvement** - Circular scenario test is added.
+
 0.7.0 (Stable)
 --------------